fix: add nebula-lint and rename class and package
