+++ src/test/java/no/unit/examplepackage/ExampleClassTest.java
@@ -1,4 +1,4 @@
-package no.unit.foo;
+package no.unit.examplepackage;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
@@ -8,13 +8,13 @@ import static org.hamcrest.core.IsNull.nullValue;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-public class FoobarTest {
+public class ExampleClassTest {
 
     @DisplayName("setField should set the field value")
     @Test
     void setFieldSetsTheFieldValue() {
         String insertedValue = "SomeValue";
-        Foobar foo = new Foobar();
+        ExampleClass foo = new ExampleClass();
         foo.setField(insertedValue);
         String actualValue = foo.getField();
         assertThat(actualValue, is(equalTo(insertedValue)));
@@ -23,7 +23,7 @@ public class FoobarTest {
     @DisplayName("getField returns null when value has not been set")
     @Test
     void getFieldSReturnsNullWhenValueHasNotBeenSet() {
-        Foobar foo = new Foobar();
+        ExampleClass foo = new ExampleClass();
         String actualValue = foo.getField();
         assertThat(actualValue, is(nullValue()));
     }
@@ -32,7 +32,7 @@ public class FoobarTest {
     @Test
     void getFieldReturnsTheSetValueWhenTheValueHasBeenSet() {
         String insertedValue = "SomeValue";
-        Foobar foo = new Foobar();
+        ExampleClass foo = new ExampleClass();
         foo.setField(insertedValue);
         String actualValue = foo.getField();
         assertThat(actualValue, is(equalTo(insertedValue)));
